MA-13971-1 [Android] HVAC: system can't show the temperature correctly when set from M4 side.
the data shown in hvac panel is Fahrenheit and the default data
send to M4 is Celsius.
If you set as below(0x41C46666 mean Celsius 24.55), the panel will shown 24.
report ac_temp 0x41C46666 1
It's not correct. The date in panel should be Fahrenheit 76.

the mValue is not correct, this patch pass correct data userTemperature.

Change-Id: I8737e64931327528b1bca7456d88e3d82410bebd
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>