MA-14244 HVAC: The HVAC Panel interface will flash back and the serial input command will be invalid.(100%)
When HVAC is in "power off" state, setFanSpeed and setFanDirection
will get a StatusCode::NOT_AVAILABLE with an IllegalStateException,
which causes this app crash. This is Vehicle hal normal behaviour.

Now catch IllegalStateException in these cases to avoid app crash.
The FanSpeed and FanDirection UI will change, but the real value
will not change.

Change-Id: I47a3ec75f8731b99d19676182d414c5e8dbf8654
Signed-off-by: Jindong <jindong.yue@nxp.com>